refactoring basketPage
