+++ src/app/App.js
@@ -5,7 +5,7 @@ import Products from "./layouts/products";
 import Login from "./layouts/login";
 import Main from "./layouts/main";
 import NavBar from "./components/ui/navBar";
-import BasketHeader from "./layouts/basketHeader";
+import Basket from "./layouts/basket";
 import AdminPage from "./components/page/adminPage/adminPage";
 // import AdminPage0 from "./components/page/adminPage/adminPage0";
 
@@ -18,7 +18,7 @@ const App = () => {
                 <Route path="/login/:type?" component={Login} />
                 {/* <Route path="/adminPage0" component={AdminPage0} /> */}
                 <Route path="/adminPage" component={AdminPage} />
-                <Route path="/basketHeader" component={BasketHeader} />
+                <Route path="/basket" component={Basket} />
                 <Route path="/" exact component={Main} />
                 <Redirect to="/" />
             </Switch>
