+++ src/app/components/page/basketCartList/index.js
@@ -1,0 +1,2 @@
+import BasketCartList from "./basketCartList";
+export default BasketCartList;
