+++ src/app/components/page/basketShopList/index.js
@@ -1,0 +1,2 @@
+import BasketShopList from "./basketShopList";
+export default BasketShopList;
