+++ src/app/layouts/basket.jsx
@@ -1,0 +1,8 @@
+import React from "react";
+import BasketForm from "../components/ui/basketForm";
+
+const Basket = () => {
+    return <BasketForm />;
+};
+
+export default Basket;
