+++ src/app/layouts/products.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useParams } from "react-router-dom";
 
 import ProductsListPage from "../components/page/productsListPage";
-import BasketShopPage from "../components/page/basket";
+import BasketShopPage from "../components/page/basketShopPage";
 
 const Products = () => {
     const params = useParams();
